Change Floats to Ints

+++ server/schema/typeDefs.js
@@ -102,8 +102,8 @@ const typeDefs = gql`
         register(email: String!, username: String!, password: String!): User
         login(identifier: String!, password: String!): User
         logout: String
-        addToCollection(userId: ID!, whiskeyId: ID!, userRating: Float, userNotes: WhiskeyNotesInput): UserCollection
-        updateReview(userId: ID!, whiskeyId: ID!, userRating: Float, userNotes: WhiskeyNotesInput): UserCollection
+        addToCollection(userId: ID!, whiskeyId: ID!, userRating: Int, userNotes: WhiskeyNotesInput): UserCollection
+        updateReview(userId: ID!, whiskeyId: ID!, userRating: Int, userNotes: WhiskeyNotesInput): UserCollection
         removeFromCollection(userId: ID!, whiskeyId: ID!): UserCollection
         addToWishlist(userId: ID!, whiskeyId: ID!): UserWishlist
         removeFromWishlist(userId: ID!, whiskeyId: ID!): UserWishlist
